+++ R/getRadarTable.R
@@ -4,8 +4,7 @@
 #' the columns appropriately
 #' @author Fabian Hertner, \email{fabian.hertner@@swiss-birdradar.com};
 #' Birgen Haest, \email{birgen.haest@@vogelwarte.ch}
-#' @param dbConnection a valid  database connection
-#' @param dbDriverChar the name of the driver.
+#' @inheritParams QUERY
 #'
 #' @return the radar table  as a data frame
 #' @export
@@ -27,11 +26,15 @@
 #' )
 #' dbConnection = RODBC::odbcDriverConnect(dsn)
 #'
-#' radarTable = getRadarTable(dbConnection, dbDriverChar)
+#' radarTable = getRadarTable(dbConnection)
 #' }
 #'
 getRadarTable = function(dbConnection, dbDriverChar) {
-  radarTable = QUERY(dbConnection, dbDriverChar, "SELECT * FROM radar")
+  if (!missing(dbDriverChar)) {
+    lifecycle::deprecate_warn("0.4.0", "getRadarTable(dbDriverChar)")
+  }
+
+  radarTable = QUERY(dbConnection, query = "SELECT * FROM radar")
 
   colnames(radarTable)[colnames(radarTable) == "radarid"] = "radarID"
   colnames(radarTable)[colnames(radarTable) == "serialno"] = "serialNo"
@@ -52,6 +55,5 @@ getRadarTable = function(dbConnection, dbDriverChar) {
   colnames(radarTable)[colnames(radarTable) == "pulselengthlong"] = "pulseLengthLong"
   colnames(radarTable)[colnames(radarTable) == "tiltangle"] = "tiltAngle"
   colnames(radarTable)[colnames(radarTable) == "transmitpower"] = "transmitPower"
-
   return(radarTable)
 }