+++ R/getVisibilityTable.R
@@ -4,11 +4,9 @@
 #' 'SQL' database
 #' @author Fabian Hertner, \email{fabian.hertner@@swiss-birdradar.com};
 #' Birgen Haest, \email{birgen.haest@@vogelwarte.ch}
-#' @param dbConnection a valid  database connection
-#' @param dbDriverChar the name of the driver. If different from 'PostgreSQL'
-#' it connects to cloud.birdradar.com
+#' @inheritParams QUERY
 #'
-#' @return A dataframe with the visibility table
+#' @return A `data.frame` with the visibility table
 #' @export
 #'
 #' @examples
@@ -29,58 +27,47 @@
 #' )
 #' dbConnection = RODBC::odbcDriverConnect(dsn)
 #'
-#' visibilityTable = getVisibilityTable(dbConnection, dbDriverChar)
+#' visibilityTable = getVisibilityTable(dbConnection)
 #' }
 #'
-getVisibilityTable = function(dbConnection, dbDriverChar) {
+getVisibilityTable = function(dbConnection, dbDriverChar){
+  if (!missing(dbDriverChar)) {
+    lifecycle::deprecate_warn("0.4.0", "getVisibilityTable(dbDriverChar)")
+  }
   # load protocol table from local MS-SQL DB
   # ============================================================================
-  if (dbDriverChar != "PostgreSQL") {
-    visibilityTable = QUERY(
-      dbConnection,
-      dbDriverChar,
-      paste0(
-        "SELECT * FROM visibility ",
-        "order by visibilityLogID asc"
-      )
-    )
-    visibilityTable_times = QUERY(dbConnection,
-      dbDriverChar,
-      paste0(
-        "SELECT blind_from, blind_to ",
-        "FROM visibility order by ",
-        "visibilityLogID asc"
-      ),
-      as.is = TRUE
-    )
-    visibilityTable$blind_from = visibilityTable_times$blind_from
-    visibilityTable$blind_to = visibilityTable_times$blind_to
+    if (class(dbConnection) != 'PqConnection'){
+      visibilityTable            = QUERY(dbConnection,
+                                         query = paste0("SELECT * FROM visibility ",
+                                                "order by visibilityLogID asc"))
+      visibilityTable_times      = QUERY(dbConnection,
+                                         query = paste0("SELECT blind_from, blind_to ",
+                                                "FROM visibility order by ",
+                                                "visibilityLogID asc"),
+                                         as.is = TRUE)
+      visibilityTable$blind_from = visibilityTable_times$blind_from
+      visibilityTable$blind_to   = visibilityTable_times$blind_to
 
-    # load protocol table from PostGreSQL
-    # ============================================================================
-  } else {
-    visibilityTable = QUERY(
-      dbConnection,
-      dbDriverChar,
-      paste0(
-        "SELECT *,blind_from::character",
-        " varying as ",
-        "blindfrom,blind_to::character",
-        " varying as blindto FROM ",
-        "visibility order by ",
-        "visibilitylogid asc"
-      )
-    )
-    visibilityTable$blind_from = visibilityTable$blindfrom
-    visibilityTable$blind_to = visibilityTable$blindto
-    visibilityTable$blindfrom = NULL
-    visibilityTable$blindto = NULL
-  }
+  # load protocol table from PostGreSQL
+  # ============================================================================
+    } else {
+      visibilityTable            = QUERY(dbConnection,
+                                       query =   paste0("SELECT *,blind_from::character",
+                                                " varying as ",
+                                                "blindfrom,blind_to::character",
+                                                " varying as blindto FROM ",
+                                                "visibility order by ",
+                                                "visibilitylogid asc"))
+      visibilityTable$blind_from = visibilityTable$blindfrom
+      visibilityTable$blind_to   =  visibilityTable$blindto
+      visibilityTable$blindfrom  = NULL
+      visibilityTable$blindto    =  NULL
+    }
 
 
-  colnames(visibilityTable)[colnames(visibilityTable) == "visibilitylogid"] =
-    "visibilityLogID"
-  colnames(visibilityTable)[colnames(visibilityTable) == "protocolid"] =
-    "protocolID"
-  return(visibilityTable)
+    colnames(visibilityTable)[colnames(visibilityTable) == "visibilitylogid"] =
+      "visibilityLogID"
+    colnames(visibilityTable)[colnames(visibilityTable) == "protocolid"]      =
+      "protocolID"
+    return(visibilityTable)
 }