+++ R/plotExploration.R
@@ -7,7 +7,7 @@
 #' observed echoes at their respective altitudes. These plots are helpful to
 #' roughly visually explore your data (and for example spot oddities).
 #' @param echoData dataframe with the echo data from the data list created by
-#' the function ‘extractDBData’ or a subset of it created by the function
+#' the function [extractDbData()] or a subset of it created by the function
 #' ‘filterEchoData’
 #' @param timeRange optional list of string vectors length 2, start and end
 #' time of the time ranges that should be plotted. The date/time format is
@@ -20,10 +20,10 @@
 #' created by the function ‘loadManualBlindTimes’. If not set, manual blind
 #' times are not shown in the plot.
 #' @param visibilityData optional dataframe with the visibility data created by
-#' the function ‘extractDBData’. If not set, visibility data are not shown in
+#' the function [extractDbData()]. If not set, visibility data are not shown in
 #' the plot.
 #' @param protocolData optional dataframe with the protocol data used to filter
-#' the echoes, created by the function ‘extractDBData’ or a subset of it created
+#' the echoes, created by the function [extractDbData()] or a subset of it created
 #' by the function ‘filterProtocolData’. If not set, periods without a protocol
 #' are not shown in the plot.
 #' @param sunriseSunset optional dataframe with sunrise/sunset, civil, and nautical