+++ R/query.R
@@ -3,8 +3,8 @@
 #' @description  Run an 'SQL' query on an already connected database
 #' @author Fabian Hertner, \email{fabian.hertner@@swiss-birdradar.com};
 #' Birgen Haest, \email{birgen.haest@@vogelwarte.ch}
-#' @param dbConnection a valid  database connection
-#' @param dbDriverChar the name of the driver
+#' @param dbConnection a valid database connection
+#' @param dbDriverChar This was the name of the driver, and is now automatically detected, therefore it should be omitted into the future.
 #' @param query an 'SQL' string with your query
 #' @param as.is If TRUE, leaves data as it is
 #'
@@ -31,16 +31,23 @@
 #'
 #' QUERY(
 #'   dbConnection = dbConnection,
-#'   dbDriverChar = dbDriverChar,
 #'   query = "Select * From collection order by row asc"
 #' )
 #' }
 #'
 QUERY <- function(dbConnection, dbDriverChar, query, as.is = FALSE) {
-  if (dbDriverChar == "PostgreSQL") {
+  if (!missing(dbDriverChar)) {
+    lifecycle::deprecate_warn("0.4.0", "QUERY(dbDriverChar)")
+  }
+  if (class(dbConnection) %in% "PqConnection") {
     t <- DBI::dbGetQuery(dbConnection, query, as.is = as.is)
   } else {
-    t <- RODBC::sqlQuery(dbConnection, query, as.is = as.is)
+    if (class(dbConnection) %in% "RODBC") {
+      t <- RODBC::sqlQuery(dbConnection, query, as.is = as.is)
+    } else {
+      stop("The `dbConnection` argument now only supports connections of the class `PqConnection` and `RODBC`.")
+    }
   }
+
   return(t)
 }