+++ NEWS.md
@@ -4,6 +4,7 @@
 - New functionality: the functions `getCollectionTable()` and `extractDbData()` now include an additional optional parameter `timeInterval` for when you want to limit the data being retrieved from the collection table (#49).
 - New functionality: the function `extractDBData()` now includes an additional optional parameter `dbHost` to set your postgreSQL database host site (#54).
 - New functionality: the function `getEchoFeatures()` now includes an additional optional parameter `echoSubset` for when you want to limit the data being retrieved from the `echo_rffeature_map` table to a certain subset of echoes only. In the function `extractDBData()`, the default is now to not extract the full `echo_rffeature_map` table anymore, but only for the echoes included in the `collection` table (which can now be subset based on a time interval). This improves performance for very large databases.
+- Simplification: All `get` functions do not require the `dbDriverChar` argument.
 
 # birdscanR 0.3.0